+++ source/libnormaliz/integer.h
@@ -189,7 +189,7 @@ mpq_class dec_fraction_to_mpq(string s);
 // there is a remainder
 
 bool int_quotient(long& Quot, const long long& Num, const long& Den);
-bool int_quotient(long long& Quot, const long& Num, const long& Den);
+bool int_quotient(long long& Quot, const long long& Num, const long long& Den);
 bool int_quotient(mpz_class& Quot, const mpz_class& Num, const mpz_class& Den);
 template<typename IntegerRet>
 bool int_quotient(IntegerRet& Quot, const nmz_float& Num, const nmz_float& Den);
